agent-service/pyproject.toml
[project]
name = "fairworkly-agent-service"
version = "0.1.0"
description = "FastAPI service hosting FairWorkly's compliance agent."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pypdf>=4.0",
    "docx2txt>=0.8",
    "transformers>=4.40",
    "huggingface-hub>=0.23",
    "torch>=2.2",
    "sentence-transformers>=2.4",
    "pytest>=7.0",
    "httpx>=0.27",
    "pydantic>=2.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
